Modified enyo.Popup to not call release() to stop capturing events unless it has previously called capture() to start
This caused problems with the context menu in Designer, where if the click that triggers the context menu also selects the component the creation of Inspector controls caused the context menu to lose focus.  There's a fix for this in enyo that we haven't gotten yet; this fix is a stop-gap until we update.