Prefer requestAnimationFrame over any vendor-prefixed version
refs #6138
Prior to d108536c in the main Enyo repo, enyo.requestAnimationFrame wrapped any vendor-prefixed version over the unprefixed version. d108536c changed the order in which versions are checked. It also adds checking for cancelAnimationFrame as well as cancelRequestAnimationFrame; Chrome is about to remove the latter so without porting these changes we’ll degrade to setTimeout.